+++ Samples/Graphics/Lights/LightsWindow.cpp
@@ -3,7 +3,7 @@
 // Distributed under the Boost Software License, Version 1.0.
 // http://www.boost.org/LICENSE_1_0.txt
 // http://www.geometrictools.com/License/Boost/LICENSE_1_0.txt
-// File Version: 3.0.0 (2016/06/19)
+// File Version: 3.0.1 (2018/10/05)
 
 #include "LightsWindow.h"
 
@@ -141,8 +141,8 @@ bool LightsWindow::OnCharPress(unsigned char key, int x, int y)
                 auto lighting = effect->GetLighting();
                 lighting->spotCutoff[0] -= 0.1f;
                 lighting->spotCutoff[0] = std::max(lighting->spotCutoff[0], 0.0f);
-                lighting->spotCutoff[1] = cos(lighting->spotCutoff[0]);
-                lighting->spotCutoff[2] = sin(lighting->spotCutoff[0]);
+                lighting->spotCutoff[1] = std::cos(lighting->spotCutoff[0]);
+                lighting->spotCutoff[2] = std::sin(lighting->spotCutoff[0]);
                 effect->UpdateLightingConstant();
             }
         }
@@ -157,8 +157,8 @@ bool LightsWindow::OnCharPress(unsigned char key, int x, int y)
                 auto lighting = effect->GetLighting();
                 lighting->spotCutoff[0] -= 0.1f;
                 lighting->spotCutoff[0] = std::min(lighting->spotCutoff[0], (float)GTE_C_HALF_PI);
-                lighting->spotCutoff[1] = cos(lighting->spotCutoff[0]);
-                lighting->spotCutoff[2] = sin(lighting->spotCutoff[0]);
+                lighting->spotCutoff[1] = std::cos(lighting->spotCutoff[0]);
+                lighting->spotCutoff[2] = std::sin(lighting->spotCutoff[0]);
                 effect->UpdateLightingConstant();
             }
         }
@@ -213,8 +213,8 @@ void LightsWindow::CreateScene()
     // Fedora 21 Linux.
     Vector4<float> darkGray{ 0.1f, 0.1f, 0.1f, 1.0f };
     Vector4<float> lightGray{ 0.75f, 0.75f, 0.75f, 1.0f };
-    float angle = 0.125f*(float)GTE_C_PI;
-    Vector4<float> lightSpotCutoff{ angle, (float)cos(angle), (float)sin(angle), 1.0f };
+    float angle = 0.125f * (float)GTE_C_PI;
+    Vector4<float> lightSpotCutoff{ angle, std::cos(angle), std::sin(angle), 1.0f };
 
     mLightWorldPosition[SVTX] = { 4.0f, 4.0f - 8.0f, 8.0f, 1.0f };
     mLightWorldPosition[SPXL] = { 4.0f, 4.0f + 8.0f, 8.0f, 1.0f };
